11/8 - 12.15pm FutureDates url okay, deleted employee backup okay, view own asset url okay.

+++ Backend 1.0/src/main/java/com/project/inventoryManagement/Controllers/EmployeeMainController.java
@@ -1,8 +1,11 @@
 package com.project.inventoryManagement.Controllers;
 
 
+import com.project.inventoryManagement.Models.DeletedEmployeesModel;
 import com.project.inventoryManagement.Models.EmployeeMainModel;
+import com.project.inventoryManagement.Repositories.DeletedEmployeesRepo;
 import com.project.inventoryManagement.Repositories.EmployeeMainRepository;
+import com.project.inventoryManagement.Service.DeletedEmployeeMapping;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -22,6 +25,13 @@ public class  EmployeeMainController {
 
     @Autowired
     private PasswordEncoder bcryptEncoder;
+
+    @Autowired
+    private DeletedEmployeeMapping deletedEmployeeMapping;
+
+    @Autowired
+    private DeletedEmployeesRepo deletedEmployeesRepo;
+
 
 
 
@@ -124,18 +134,23 @@ public class  EmployeeMainController {
 
       // Delete an employee by Nic
     @DeleteMapping(path = "/delete")
-    public boolean deleteEmployeeByNic(@RequestParam String nic){
+    public boolean deleteEmployeeByNic(@RequestParam String nic) throws Exception {
         System.out.println("delete nic received: "+ nic);
-        try {
+
+            EmployeeMainModel m1 = empMainRepo.findByNic(nic);
+
+            if(m1.getStatus().equals("Asset Manager")){
+                throw new Exception("Asset Manager cannot be deleted");
+            }
+
+            DeletedEmployeesModel dm = deletedEmployeeMapping.mapToDeletedEmployeeModel(m1);
+            deletedEmployeesRepo.save(dm);
+            System.out.println("Deleted employee saved to deleteed table\nNow deleting...");
+
             empMainRepo.deleteByNic(nic);
             System.out.println("Deleted");
             return true;
-        } catch(Exception e) {
-            System.out.println("Exception display as : " + e);
-//            EmployeeMainModel mm = empMainRepo.findByNic(nic);
-//            for(asset : mm.getAssign(). )
-            return false;
-        }
+
 
     }
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
